Optimisation.
There's a fine line between requiring an application to be aware of how things block, and being usable.  This change reintroduces the repeated non-blocking reads optimisation in a more natural fashion, taking advantage of how all results of socket calls are passed unbuffered between caller and underlying socket.

An interesting experiment for a user of this module is to enable the print statements relating to "handle_read:CALL" and "_recv:FORCE-CHANNEL-CHANGE" and observe how the data comes in on their socket.

Future work:
- Consider using the standard library logging module instead of prints.
- Consider bounding time spent in poll calls based on actual time passed, rather than whichever of the limits on amount of data read and number of reads is hit first.