Bringing back internal CCP modifications to this module.
It has been shown that the way that calls are proxied over a channel to the inner asyncore select loop events should not be restricted to being done consecutively when the same argument tuple is used.  The primary change here removes that limitation.

The secondary change is to use a deques instead of lists.